pyproject.toml
[project]
name = "product-graph"
version = "0.1.0"
description = ""
authors = [
    {name = "bpatel347",email = "bpatel347@gatech.edu"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "umap-learn (>=0.5.9.post2,<0.6.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
